Add Foyer management and update Universite
This commit introduces the Foyer module which includes the feature to add, list, and modify Foyers. Several updates have been made in the existing Universite as well. Changes in the Universite include the removal of the mandatory "universite" property and the addition of a UniversiteService class for interacting with a backend API. A few updates have also been made in the project title of the index.html file and a new file, universite.service.spec.ts has been added for testing the newly created service. Similarly, a completely new routing has been introduced for Foyer management. This modification was necessary in order to scale the application and manage user roles more effectively.